Always compare the entire key in HFSX binary comparisons
The memcmp in hfslib_compare_catalog_keys_bc would, on occasion, only
check roughly half of the provided key. This led to lookups on
"dir_205123" returning the entry for another catalog entry,
"dir_205789", which shares a prefix.

This failure's reproduceability is of course contingent upon the order
of the B-tree nodes in the filesystem.

Signed-off-by: Dustin L. Howett <dustin@howett.net>