Remove ublio nblocks option and fix its blocksize to 512
Ublio does not handle overreads gracefully, which may happen
if the block size for reading is much larger than the volume's
block size and the volume is small.

The volume's internal block size is not yet known at the time
we initialize ublio, so we use a safe minimum value.

Builds without ublio or using the noublio option are unaffected.

This value can still be overridden with the blksize= option.