hfsfuse: Use unprefixed FUSE header for compatibility
More systems support this out of the box, and those that don't can
be handled by adding to the include path, while the inverse is not true.

This is also the expected header when using pkg-config to discover FUSE,
though we don't do this currently, and not all systems that break from
this convention ship with a fuse.pc.