libhfs: Check record count when reading user data from header nodes
Right now libhfs assumes that if the caller requested the user
data or map records for a given B-tree they'll also provide the
corresponding record sizes. This amends the fix in
c7f8db2d33c67eb6992be73ddabe880626ee87b2 to make sure of this.

This functionality is not actually used by hfsfuse/hfsdump, so
this change is only relevant if there are other users of our
bundled libhfs.