Makefile: fix feature detection with -pedantic
Within the cccheck call, `({ statement })` was initially
used for consistency as the parens ensure that any commas in the
statement will be interpreted by make as part of the argument.
However, in C this results in a block within a statement, which
is nonstandard.

We retain the additional block braces although this change makes
them purely cosmetic.