Translate IFMT flags to system native values
These higher bits of st_mode are not standardized and so the value of
e.g. S_IFDIR may differ between the filesystem and the host system.

TN1150 does define specific values for these flags, so we use the
provided definitions to convert to the system's corresponding S_IF*
values.