hfsfuse: prevent com.apple.ResourceFork xattr lookup for folder records
Fixes a potential uninitialized read when retrieving this attribute's
length. Further attempts to read from these were prevented by libhfs and
resulted in an EIO response to getxattr; this now results in ENOATTR
as expected.

We also no longer present this attribute for files with resource forks
of length larger than INT_MAX as extended attributes in FUSE are limited
to this, fixing a potential overflow of the returned size.