Fix check for empty com.apple.FinderInfo xattrs
This is missing an offset to the file record's user_info, but this would
not be portable anyway due to the possibility of padding so instead just
check if the serialized buffer is empty.

The effect is that previously this xattr would always be listed despite
its value being all zeros.