hfsfuse: add explicit checks for read size/offset
This is a bit redundant as no existing FUSE implementation issues reads
beyond the size reported by getattr for a given file, and if they did
they would only receive the zero padding within the file's last block.

However this generalizes the previous special case check for zero length
files, and means that we don't have to trust the FUSE implementation to
request exact values.