Make better use of GNU Make builtins
Build commands now use implicit rules if they can or follow the same
conventions when defined explicitly. This means that for example the
value of LDLIBS if set will be used when linking, whereas before
any additional libraries would have needed to be included in the CFLAGS.

Also removes redundant defaults for CC, AR, and RM which gmake
implicitly defines to the same values.