hfsfuse: ensure readlink result is NUL terminated
This is required by libfuse, which needs a valid C string to determine
the length of the returned link destination.

Fixes a potential read and disclosure of uninitialized memory of up to
PATH_MAX bytes when reading link destinations on a mounted filesystem.
Bytes following PATH_MAX are terminated by FUSE.

Thanks to @sbond75 for finding this.