libhfs: fix memory corruption in certain extent record lookups
This could occur during extent lookup when a thread record with a name
longer than 123 characters preceded the target file in the same B-tree
node, which resulted in an out of bounds write to the stack that
partially corrupted the neighboring input search key.

In practice this resulted in libhfs being unable to find extent records
for the affected files, resulting in an error when attempting to read
them.