hfsfuse: make extended attribute namespace configurable
Changes this to a build-time make parameter XATTR_NAMESPACE which can
be used to select other namespaces than the OS default.

Allows the platform dependent checks for this to be moved out to the
build system and provides a means for accessing symlink extended
attributes on Linux, which can't be read under the user namespace.