hfs_stat: saturate nlink stat field on overflow
On some platforms (including macOS) nlink_t is a smaller type than the
folder record's valence. This represents the number of entries in a
directory, so it's quite possible to overflow this on a normal volume.

hfsfuse now sets st_nlink to the max possible value in this case,
consistent with other drivers.