hfsuser: extract logic for full-length reads
`read`/`pread` only typically guarantee complete reads of the given
length for regular files. This adds hfs_preadall which either completes
the full read of size `nbyte` or fails with errno set.

Supercedes the existing logic for this in hfs_read_pread and fixes the
possibility of short reads on trailing blocks.