hfsuser: don't use MAXBYTECOUNTREAD for read block size on macOS
What libhfsuser is looking for with these ioctls is the best size and
alignment for reads rather than how much to read at a time. The latter
isn't necessary as hfs_read_pread always reads as many blocks as it can
contiguously.

In particular the value of DKIOCGETMAXBYTECOUNTREAD may be up to several
megabytes, which has large affects on the performance of small reads
when aligned to.