hfsfuse: don't reuse stat structs when filling readdir entries
Directories and files may fill different fields of this. Under the FUSE
version used by hfsfuse only the st_mode and st_ino fields of this are
used. These are always populated by hfs_stat, so not an issue currently,
but this will be needed for readdirplus support in the future.