libhfs: remove redundant wrapper for retrieving callbacks
libhfs already makes the global callbacks extern, making this accessor
(an hfsfuse addition) not a useful difference over upstream as it does
not add any meaningful encapsulation to these.

If these are made internal to libhfs at some point this would call for a
different interface than the one being removed here in any case.